Fix calculation bug

+++ src/OpenLoco/src/Economy/Economy.cpp
@@ -129,6 +129,6 @@ namespace OpenLoco::Economy
 
     currency32_t getInflationAdjustedCost(int16_t costFactor, uint8_t costIndex, uint8_t divisor)
     {
-        return costFactor * static_cast<int32_t>(currencyMultiplicationFactors()[costIndex]) / (1 << divisor);
+        return costFactor * static_cast<int64_t>(currencyMultiplicationFactors()[costIndex]) / (1ULL << divisor);
     }
 }
